pyproject.toml
[tool.poetry]
name = "snapbot"
version = "0.1.0"
description = ""
authors = ["kish <dhruvsharma3412@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
discord = "^2.3.2"
python-dotenv = "^1.0.1"
ruff = "^0.5.1"
cryptography = "^42.0.8"
motor = "^3.5.0"
toml = "^0.10.2"
requests = "^2.32.3"
reactionmenu = "^3.1.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
